+++ Dockerfile
@@ -1,0 +1,31 @@
+# Use the official Go image as a builder
+FROM golang:1.21-alpine AS builder
+
+# Set the working directory
+WORKDIR /app
+
+# Copy go.mod and go.sum files
+COPY go.mod go.sum ./
+
+# Download dependencies
+RUN go mod download
+
+# Copy the source code
+COPY . .
+
+# Build the application
+RUN CGO_ENABLED=0 GOOS=linux go build -o /app/CloudSnitch
+
+# Use a minimal alpine image for the final container
+FROM alpine:latest
+
+# Install ca-certificates for HTTPS requests
+RUN apk --no-cache add ca-certificates
+
+WORKDIR /app
+
+# Copy the binary from builder
+COPY --from=builder /app/CloudSnitch .
+
+# Set the entrypoint
+ENTRYPOINT ["/app/CloudSnitch"]
